Legal_Assist_ML/requirements.txt
flask
flask-cors
pymongo
python-dotenv
requests
gunicorn
